Prevent spamming terminal with 'Not realizable'
When the ODE solver encountered an unrealizable situation, it sent the string 'Not realizeable' to the terminal each time the local time step was adjusted.  This spammed the terminal/logs and was somewhat uniformative to new users and has been accidnetally considered a failed run, rather than part of the ODE solution process.  The string is now only output the first time the local time step is adjusted and has been reworded to hopefully reduce confusion.