DEV: Improved resizing of momentSet and univariateMomentSet.
Resizing a momentSet or a multivariateMomentSet requires a
labelListList of moment orders, since, in case the moment set
is expanded there is no obvious criterion to extend or adapt
the existing list of moment orders.

Resizing a univariateMomentSet does not require providing the
moment orders since they are automatically generated.
Additionally, when a univariateMomentSet is resized, the
list of zeta_k values and, if support is [0, 1], the list of
canonical moments are resized accordingly (nMoments - 1).
The lists of alpha and beta coefficients of the recurrence
relationship are also adapted to maintain consistency.

Finally, the member nMoments_ has been removed, and the
nMoments() method now calls (*this).size(), which returns
the size of the underlying scalarList, to simplify operations
when using setSize() and resize().

The code in momentSet, univariateMomentSet and multivariateMomentSet
has been updated to replace nMoments_ appropriately.